requirements.txt
beautifulsoup4==4.8.1
gunicorn
django
line-bot-sdk
psycopg2-binary